requirements.txt
undetected-chromedriver
webdriver-manager
python-decouple
tqdm
colorama
PySimpleGUI
